haqiqa_ai_part/requirements.txt
flask
mysql-connector-python
minio
openai-whisper
python-dotenv
langchain   
langchain-community   
langchain-ollama      
langchain-chroma
chromadb       
sentence-transformers